# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/medicines/{id}": {
      "get": {
        "operationId": "MedicinesController_getDetail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Medicine id in the selected source",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "required": false,
            "in": "query",
            "description": "Knowledge source selector.",
            "schema": {
              "default": "drugbank",
              "type": "string",
              "enum": ["drugbank", "cn"]
            }
          },
          {
            "name": "x-bypass-cache",
            "required": false,
            "in": "header",
            "description": "Set to true/1/no-cache to bypass medicines read cache for this request only.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedicineDetailResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get medicine detail from a selected knowledge source",
        "tags": ["Medicines"]
      }
    },
...
    "/api/v1/user/medicine-dose-logs": {
      "get": {
        "operationId": "MedicineDoseLogsController_list_v1",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-06-04",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoseLogListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List dose logs for a date",
        "tags": ["Medicine Dose Logs"]
      },
      "post": {
        "operationId": "MedicineDoseLogsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDoseLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoseLogResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a dose log",
        "tags": ["Medicine Dose Logs"]
      }
    },
    "/api/v1/user/medicine-dose-logs/{id}": {
      "patch": {
        "operationId": "MedicineDoseLogsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDoseLogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoseLogResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a dose log",
        "tags": ["Medicine Dose Logs"]
      },
      "delete": {
        "operationId": "MedicineDoseLogsController_delete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft-delete a dose log",
        "tags": ["Medicine Dose Logs"]
      }
    },
...
    "/api/v1/user/today-analysis/generate": {
      "post": {
        "operationId": "TodayAnalysisController_generate_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateTodayAnalysisDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayAnalysisResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Generate authenticated user today AI analysis",
        "tags": ["Today Analysis"]
      }
    },
    "/api/v1/user/today-analysis/generate/stream": {
      "post": {
        "operationId": "TodayAnalysisController_generateStream_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateTodayAnalysisDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayAnalysisStreamResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Stream authenticated user today AI analysis generation",
        "tags": ["Today Analysis"]
      }
    },
    "/api/v1/user/settings": {
      "get": {
        "operationId": "UserSettingsController_getSettings_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get authenticated user settings",
        "tags": ["User Settings"]
      },
      "patch": {
        "operationId": "UserSettingsController_updateSettings_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update authenticated user settings",
        "tags": ["User Settings"]
      }
    },
...
  },
  "info": {
    "title": "Lucent API",
    "description": "Lucent 后端 API 文档",
    "version": "1.0"
  },
  "components": {
    "schemas": {
...
      "DrugbankMedicineDetailDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "example": "drugbank"
          },
          "drugType": {
            "type": "object",
            "nullable": true,
            "example": "small molecule"
          },
          "state": {
            "type": "object",
            "nullable": true,
            "example": "solid"
          },
          "description": {
            "type": "object",
            "nullable": true,
            "example": "A non-steroidal anti-inflammatory drug."
          },
          "indication": {
            "type": "object",
            "nullable": true,
            "example": "Used for pain, fever, and inflammation."
          },
          "mechanismOfAction": {
            "type": "object",
            "nullable": true
          },
          "pharmacodynamics": {
            "type": "object",
            "nullable": true
          },
          "toxicity": {
            "type": "object",
            "nullable": true
          },
          "metabolism": {
            "type": "object",
            "nullable": true
          },
          "absorption": {
            "type": "object",
            "nullable": true
          },
          "halfLife": {
            "type": "object",
            "nullable": true
          },
          "proteinBinding": {
            "type": "object",
            "nullable": true
          },
          "routeOfElimination": {
            "type": "object",
            "nullable": true
          },
          "volumeOfDistribution": {
            "type": "object",
            "nullable": true
          },
          "clearance": {
            "type": "object",
            "nullable": true
          },
          "groups": {
            "example": ["approved", "small molecule"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "categories": {
            "example": ["Anti-inflammatory Agents"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "atcCodes": {
            "example": ["M01AE01"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "synonyms": {
            "example": ["Ibuprofen"],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "foodInteractions": {
            "example": ["Avoid taking with alcohol."],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "drugInteractions": {
            "type": "object",
            "description": "Raw source interaction payload.",
            "nullable": true,
            "additionalProperties": true
          },
          "externalIdentifiers": {
            "type": "object",
            "description": "Raw source external identifier payload.",
            "nullable": true,
            "additionalProperties": true
          },
          "externalLinks": {
            "type": "object",
            "description": "Raw source external link payload.",
            "nullable": true,
            "additionalProperties": true
          }
        },
        "required": [
          "kind",
          "groups",
          "categories",
          "atcCodes",
          "synonyms",
          "foodInteractions"
        ]
      },
      "CnMedicineDetailDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "example": "cnProduct"
          },
          "approvalNumber": {
            "type": "object",
            "nullable": true,
            "example": "国药准字H10900089"
          },
          "manufacturer": {
            "type": "object",
            "nullable": true,
            "example": "某某制药"
          },
          "packageSpec": {
            "type": "object",
            "nullable": true,
            "example": "0.3g*10粒"
          },
          "brandName": {
            "type": "object",
            "nullable": true,
            "example": "布洛芬"
          },
          "ingredients": {
            "type": "object",
            "nullable": true
          },
          "properties": {
            "type": "object",
            "nullable": true
          },
          "indications": {
            "type": "object",
            "nullable": true
          },
          "dosage": {
            "type": "object",
            "nullable": true
          },
          "adverseReactions": {
            "type": "object",
            "nullable": true
          },
          "contraindications": {
            "type": "object",
            "nullable": true
          },
          "precautions": {
            "type": "object",
            "nullable": true
          },
          "pediatricUse": {
            "type": "object",
            "nullable": true
          },
          "geriatricUse": {
            "type": "object",
            "nullable": true
          },
          "pregnancyLactation": {
            "type": "object",
            "nullable": true
          },
          "pharmacologyToxicology": {
            "type": "object",
            "nullable": true
          },
          "drugInteractions": {
            "type": "object",
            "nullable": true
          },
          "pharmacokinetics": {
            "type": "object",
            "nullable": true
          },
          "overdose": {
            "type": "object",
            "nullable": true
          },
          "storage": {
            "type": "object",
            "nullable": true
          },
          "validityPeriod": {
            "type": "object",
            "nullable": true
          },
          "barcode": {
            "type": "object",
            "nullable": true
          },
          "nationalDrugCode": {
            "type": "object",
            "nullable": true
          },
          "sourceUrl": {
            "type": "object",
            "nullable": true
          },
          "imageUrl": {
            "type": "object",
            "nullable": true
          }
        },
        "required": ["kind"]
      },
...
      "MedicineDetailDataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "DB01050"
          },
          "source": {
            "type": "string",
            "enum": ["drugbank", "cn"],
            "example": "drugbank"
          },
          "name": {
            "type": "string",
            "example": "Ibuprofen"
          },
          "subtitle": {
            "type": "object",
            "example": "CAS 15687-27-1",
            "nullable": true
          },
          "detail": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DrugbankMedicineDetailDto"
              },
              {
                "$ref": "#/components/schemas/CnMedicineDetailDto"
              }
            ]
          }
        },
        "required": ["id", "source", "name", "subtitle", "detail"]
      },
      "MedicineDetailResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/MedicineDetailDataDto"
          }
        },
        "required": ["code", "message", "data"]
      },
...
      "DoseLogStatus": {
        "type": "string",
        "enum": ["taken", "skipped", "missed", "planned"]
      },
      "DoseLogItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Dose log id."
          },
          "currentMedicineId": {
            "type": "object",
            "description": "Linked current medicine id."
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DoseLogStatus"
              }
            ]
          },
          "scheduledFor": {
            "type": "string",
            "description": "Scheduled date in YYYY-MM-DD format.",
            "example": "2026-06-04"
          },
          "doseText": {
            "type": "object",
            "description": "Dose text."
          },
          "note": {
            "type": "object",
            "description": "Free-text note."
          },
          "source": {
            "type": "object",
            "description": "Source.",
            "example": "manual"
          },
          "createdAt": {
            "type": "string",
            "description": "Created at (ISO 8601)."
          },
          "updatedAt": {
            "type": "string",
            "description": "Updated at (ISO 8601)."
          }
        },
        "required": ["id", "status", "scheduledFor", "createdAt", "updatedAt"]
      },
      "DoseLogListDataDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DoseLogItemDto"
            }
          }
        },
        "required": ["items"]
      },
      "DoseLogListResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/DoseLogListDataDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "CreateDoseLogDto": {
        "type": "object",
        "properties": {
          "currentMedicineId": {
            "type": "string",
            "description": "Linked current medicine id."
          },
          "status": {
            "example": "taken",
            "allOf": [
              {
                "$ref": "#/components/schemas/DoseLogStatus"
              }
            ]
          },
          "scheduledFor": {
            "type": "string",
            "description": "Scheduled date YYYY-MM-DD.",
            "example": "2026-06-04"
          },
          "doseText": {
            "type": "string",
            "description": "Dose text."
          },
          "note": {
            "type": "string",
            "description": "Free-text note."
          }
        },
        "required": ["status", "scheduledFor"]
      },
      "DoseLogResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/DoseLogItemDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "UpdateDoseLogDto": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DoseLogStatus"
              }
            ]
          },
          "doseText": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
      "GenerateTodayAnalysisDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "description": "Target date in YYYY-MM-DD format. Defaults to backend current day when omitted.",
            "example": "2026-06-12"
          }
        }
      },
      "TodayAnalysisBulletDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": ["medication", "hydration", "sleep", "general"]
          },
          "text": {
            "type": "string"
          }
        },
        "required": ["kind", "text"]
      },
      "TodayAnalysisDataDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "generatedAt": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "bullets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodayAnalysisBulletDto"
            }
          },
          "actionLabel": {
            "type": "string"
          },
          "confidenceNote": {
            "type": "string"
          }
        },
        "required": [
          "date",
          "generatedAt",
          "summary",
          "bullets",
          "actionLabel",
          "confidenceNote"
        ]
      },
      "TodayAnalysisResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/TodayAnalysisDataDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "TodayAnalysisStreamResultDto": {
        "type": "object",
        "properties": {
          "event": {
            "type": "string",
            "enum": ["summary", "result", "error", "done"]
          },
          "data": {
            "type": "object",
            "additionalProperties": true,
            "description": "SSE payload object. event=summary => { summary }, event=result => TodayAnalysisDataDto-like object, event=error => { message, code?, statusCode? }, event=done => {}."
          }
        },
        "required": ["event", "data"]
      },
      "UserSettingsDataDto": {
        "type": "object",
        "properties": {
          "aiSummariesEnabled": {
            "type": "boolean",
            "description": "Allow AI-generated summaries and advice."
          },
          "dataSharingConsent": {
            "type": "boolean",
            "description": "Consent to share anonymized data for research."
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO-8601 timestamp of last update.",
            "nullable": true
          }
        },
        "required": ["aiSummariesEnabled", "dataSharingConsent", "updatedAt"]
      },
      "UserSettingsResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "Result code.",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "Message.",
            "example": ""
          },
          "data": {
            "$ref": "#/components/schemas/UserSettingsDataDto"
          }
        },
        "required": ["code", "message", "data"]
      },
      "UpdateUserSettingsDto": {
        "type": "object",
        "properties": {
          "aiSummariesEnabled": {
            "type": "boolean",
            "description": "Allow AI-generated summaries and advice."
          },
          "dataSharingConsent": {
            "type": "boolean",
            "description": "Consent to share anonymized data for research."
          }
        }
      },
...
    }
  }
}